Customer
Okay so
I know we agreed that I would do manager but I think you should. I don't understand the whole branch idea and it's confusing me so I'm pretty sure that whatever I do with that, it's probably not going to be what you wanted.
So I'm doing customer instead. I did the display Menu page, browsing restaurants, adding to the cart, quantity of items, etc.
I did not know how to add categories to the items, it seems complicated.
I also changed some of the sql (url links for images are 15 which is too short, changed it to 40. Also price is int, changed it to float). I also manually added some sample data to restaurants and items so you can test the pages. The new sql is exported and is in the sql folder of the project.
I'll work on the view cart, past orders, create order, and rate comment tomorrow.
Cheers,