Cart Done (sort of)
Okay, so I had to make some changes to the sql to accomadate a more informative restaurant list and menu.
The changes are:
On the restaurants table, I added minimum order, delivery type (P for only pickup, PD for pickup and delivery offered), average preparation/delivery time, and delivery charge (Yes I know the P rows shouldnt have delivery charge but it can just be kept as 0).
In the cart page you can see the quantities and costs of each added item and delete whatever you don't like. At the bottom is the order total, delivery type to choose, and proceed to checkout. Check the pages with the new sql to see how the menu and everything looks.
I'll do checkout today!